backend/requirements.txt
fastapi
uvicorn
PyMuPDF
openai
pydantic
python-dotenv
